Split files
Separated the UI from the autoclicker.
